合并process 和 hub
