+++ yc-hub/pom.xml
@@ -16,6 +16,7 @@
 
     <properties>
         <java.version>17</java.version>
+        <camunda.version>7.21.0</camunda.version>
     </properties>
 
     <dependencies>
@@ -82,9 +83,20 @@
         </dependency>
 
         <dependency>
-            <groupId>com.didiglobal.turbo</groupId>
-            <artifactId>engine</artifactId>
-            <version>1.0.0</version>
+            <groupId>org.camunda.bpm.springboot</groupId>
+            <artifactId>camunda-bpm-spring-boot-starter-webapp</artifactId>
+            <version>${camunda.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.camunda.bpm.springboot</groupId>
+            <artifactId>camunda-bpm-spring-boot-starter-rest</artifactId>
+            <version>${camunda.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.camunda.community</groupId>
+            <artifactId>camunda-engine-rest-client-complete-springboot-starter</artifactId>
+            <version>7.20.4</version>
         </dependency>
 
         <dependency>
@@ -96,6 +108,18 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.alibaba</groupId>
+            <artifactId>fastjson</artifactId>
+            <version>2.0.48</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>cn.hutool</groupId>
+            <artifactId>hutool-all</artifactId>
+            <version>5.8.27</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
