+++ yc-hub/src/main/java/com/github/leapbound/yc/hub/handler/ServiceTaskSubService.java
@@ -1,4 +1,4 @@
-package com.github.leapbound.yc.camunda.handler;
+package com.github.leapbound.yc.hub.handler;
 
 
 import com.alibaba.fastjson.JSONObject;
