+++ yc-hub/src/main/java/com/github/leapbound/yc/hub/service/impl/camunda/ServiceTaskSubServiceImpl.java
@@ -1,7 +1,7 @@
 package com.github.leapbound.yc.hub.service.impl.camunda;
 
 import com.alibaba.fastjson.JSONObject;
-import com.github.leapbound.yc.camunda.handler.ServiceTaskSubService;
+import com.github.leapbound.yc.hub.handler.ServiceTaskSubService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,6 @@ import org.springframework.stereotype.Service;
 @Service
 public class ServiceTaskSubServiceImpl implements ServiceTaskSubService {
 
-    @Override
     public JSONObject execute(String method, String arguments) throws Exception {
         log.info("camunda sub service {} {}", method, arguments);
         return new JSONObject();
