+++ yc-hub/src/main/java/com/github/leapbound/yc/hub/service/impl/function/common/ProcessFunction.java
@@ -4,7 +4,7 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.github.leapbound.sdk.llm.chat.func.MyFunctionCall;
 import com.github.leapbound.yc.hub.model.FunctionExecResultDto;
-import com.github.leapbound.yc.hub.service.YcProcessService;
+import com.github.leapbound.yc.hub.service.ProcessService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -22,7 +22,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class ProcessFunction {
 
-    private final YcProcessService processService;
+    private final ProcessService processService;
 
     public FunctionExecResultDto exec(String botId, String accountId, MyFunctionCall functionCall) {
         Map<String, Object> startFormVariables = new HashMap<>();
