--- yc-process/README.md
@@ -1,27 +1,0 @@
-<h2 align="center">yucong</h2>
-
-<h4 align=center>*An Epoch-marking Intelligent Order Service System*</h4>
-
-****
-
-## [English Docs](https://leapbound.github.io/)
-
-## Contact
-
-We recommend you to contact us using [issues](https://github.com/LeapBound/yucong/issues).
-
-## Details
-
-A chat system that can be used in multiple scenarios. For example, orders and customer service systems. At the beginning, you need to configure some of your own environment, configure your own process, and cooperate with the function-calling of the process to complete the entire system.
-
-- yc-action-server (*Function-calling implementation and script execution*)
-- yc-hub (*integrate with openai*)
-- yc-process (*integrate with camunda*)
-
-Demo is coming.
-
-## Contributors
-
-Thanks to all the contributors!
-
-[![contributors](https://contrib.rocks/image?repo=LeapBound/yucong)](https://github.com/LeapBound/yucong/graphs/contributors)
