--- yc-process/src/main/java/com/github/leapbound/yc/camunda/model/bo/ProcessVariablesRequest.java
@@ -1,23 +1,0 @@
-package com.github.leapbound.yc.camunda.model.bo;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-/**
- * @author yamath
- * @since 2023/11/22 17:57
- */
-@Data
-@NoArgsConstructor
-public class ProcessVariablesRequest {
-
-    private String processInstanceId;
-
-    private String key;
-
-    private String businessKey;
-
-    private Map<String, Object> inputVariables;
-}
