--- yc-process/src/main/java/com/github/leapbound/yc/camunda/model/vo/ProcessStepInputForm.java
@@ -1,21 +1,0 @@
-package com.github.leapbound.yc.camunda.model.vo;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * @author yamath
- * @since 2023/11/17 11:12
- */
-@Data
-@NoArgsConstructor
-public class ProcessStepInputForm {
-
-    private String id;
-
-    private String label;
-
-    private String type;
-
-    private String defaultValue;
-}
