--- yc-process/src/main/java/com/github/leapbound/yc/camunda/model/vo/ProcessVariables.java
@@ -1,19 +1,0 @@
-package com.github.leapbound.yc.camunda.model.vo;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-/**
- * @author yamath
- * @since 2023/11/22 17:59
- */
-@Data
-@NoArgsConstructor
-public class ProcessVariables {
-
-    private String processInstanceId;
-
-    private Map<String, Object> processVariables;
-}
