--- yc-process/src/main/java/com/github/leapbound/yc/camunda/service/CamundaProcessService.java
@@ -1,15 +1,0 @@
-package com.github.leapbound.yc.camunda.service;
-
-import java.util.Map;
-
-/**
- * @author yamath
- * @date 2023/11/16 10:55
- */
-public interface CamundaProcessService {
-
-    void getProcessDefinitionList();
-    void getProcessDefinitionByKey(String key);
-    String startProcessInstanceByKey(String key, Map<String, Object> variables);
-
-}
