--- yc-process/src/test/java/com/github/leapbound/yc/camunda/CommonTestApplication.java
@@ -1,15 +1,0 @@
-package com.github.leapbound.yc.camunda;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * @author tangxu
- * @date 2023/11/10 13:49
- */
-@SpringBootApplication
-public class CommonTestApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(CommonTestApplication.class, args);
-    }
-}
